% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPviz.R
\name{EXPviz}
\alias{EXPviz}
\title{EXPviz}
\usage{
EXPviz(countdata, metadata)
}
\arguments{
\item{countdata}{path to the count data csv file}

\item{metadata}{path to the metadata csv file}
}
\description{
EXPviz
}
